Generate_question/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.3.0
pydantic==2.6.1
python-multipart==0.0.9
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.0
gensim>=4.1.0
geopy>=2.2.0 